+++ src/service/utils/DurationAdapter.java
@@ -1,0 +1,35 @@
+package service.utils;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+
+import java.io.IOException;
+import java.time.Duration;
+
+public class DurationAdapter extends TypeAdapter<Duration> {
+
+    @Override
+    public void write(JsonWriter writer, Duration value) throws IOException {
+        if (value == null) {
+            writer.nullValue();
+        } else {
+            writer.value(value.toString());
+        }
+    }
+
+    @Override
+    public Duration read(JsonReader reader) throws IOException {
+        if (reader.peek() == com.google.gson.stream.JsonToken.NULL) {
+            reader.nextNull();
+            return null;
+        }
+        String durationString = reader.nextString();
+        try {
+            return Duration.parse(durationString);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
+}
