+++ src/service/utils/LocalDateTimeAdapter.java
@@ -1,0 +1,31 @@
+package service.utils;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonWriter;
+
+public class LocalDateTimeAdapter extends TypeAdapter<LocalDateTime> {
+    private final DateTimeFormatter timeFormatter = DateTimeFormatter.ofPattern("dd-MM-yyyy HH:mm:ss");
+
+    @Override
+    public void write(JsonWriter writer, LocalDateTime value) throws IOException {
+        if (value == null) {
+            writer.nullValue();
+        } else {
+            writer.value(value.format(timeFormatter));
+        }
+    }
+
+    @Override
+    public LocalDateTime read(JsonReader reader) throws IOException {
+        if (reader.peek() == com.google.gson.stream.JsonToken.NULL) { // Правильная проверка на null
+            reader.nextNull();
+            return null;
+        }
+        String timeString = reader.nextString();
+        return LocalDateTime.parse(timeString, timeFormatter);
+    }
+}
