+++ test/service/HttpTaskManagerTasksTest.java
@@ -1,0 +1,75 @@
+package service;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.AfterEach;
+
+import service.utils.DurationAdapter;
+import service.utils.LocalDateTimeAdapter;
+import task.Status;
+import task.Task;
+
+import java.io.IOException;
+import java.net.URI;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+import java.time.Duration;
+import java.time.LocalDateTime;
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+public class HttpTaskManagerTasksTest {
+
+    // создаём экземпляр InMemoryTaskManager
+    TaskManager manager = new InMemoryTaskManager();
+    int port = 8080;
+    // передаём его в качестве аргумента в конструктор HttpTaskServer
+    HttpTaskServer taskServer = new HttpTaskServer(port, manager);
+        Gson gson = Managers.getGson();
+
+    public HttpTaskManagerTasksTest() throws IOException {
+    }
+
+    @BeforeEach
+    public void setUp() throws IOException {
+        manager.removeTasks();
+        manager.removeEpics();
+        manager.removeSubtasks();
+        taskServer.start();
+    }
+
+//    @AfterEach
+//    public void shutDown() {
+//        taskServer.stop();
+//    }
+
+    @Test
+    public void testAddTask() throws IOException, InterruptedException {
+        // создаём задачу
+        Task task = new Task("Test 2", "Testing task 2",
+                Status.NEW, LocalDateTime.now(), Duration.ofMinutes(5));
+        // конвертируем её в JSON
+        String taskJson = gson.toJson(task);
+
+        // создаём HTTP-клиент и запрос
+        HttpClient client = HttpClient.newHttpClient();
+        URI url = URI.create("http://localhost:8080/tasks/task");
+        HttpRequest request = HttpRequest.newBuilder().uri(url).POST(HttpRequest.BodyPublishers.ofString(taskJson)).build();
+        // вызываем рест, отвечающий за создание задач
+        HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
+        // проверяем код ответа
+        assertEquals(201, response.statusCode());
+
+        // проверяем, что создалась одна задача с корректным именем
+        List<Task> tasksFromManager = manager.getAllTask().stream().toList();
+
+        assertNotNull(tasksFromManager, "Задачи не возвращаются");
+        assertEquals(1, tasksFromManager.size(), "Некорректное количество задач");
+        assertEquals("Test 2", tasksFromManager.get(0).getName(), "Некорректное имя задачи");
+    }
+}
